chatbot_api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langgraph
langchain-google-vertexai
langchain-qdrant
pyodbc
